requirements.txt
alembic
flask
flask-assets
flask-login
flask-wtf
requests
pyyaml
